Fix review comment about viewController dependencies (#924)
