+++ compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/window/ComposeBridge.uikit.kt
@@ -47,77 +47,3 @@ internal interface ComposeBridge {
     @Composable
     fun ProvideRootCompositionLocals(content: @Composable () -> Unit)
 }
-
-@OptIn(InternalComposeApi::class)
-internal fun createComposeBridge(
-    configuration: ComposeUIViewControllerConfiguration,
-    content: @Composable () -> Unit,
-) = object : ComposeBridge {
-    override val layers: MutableList<ComposeSceneLayerBridge> = mutableListOf()
-    override val configuration = configuration
-    override val layoutDirection get() = getLayoutDirection()
-    private val composeSceneBridges: MutableList<ComposeSceneBridge> = mutableListOf()
-
-    /*
-     * Initial value is arbitrarily chosen to avoid propagating invalid value logic
-     * It's never the case in real usage scenario to reflect that in type system
-     */
-    val interfaceOrientationState: MutableState<InterfaceOrientation> = mutableStateOf(
-        InterfaceOrientation.Portrait
-    )
-    val systemThemeState: MutableState<SystemTheme> = mutableStateOf(SystemTheme.Unknown)
-
-    override val focusStack: FocusStack<UIView> = FocusStackImpl()
-
-    @Composable
-    override fun ProvideRootCompositionLocals(content: @Composable () -> Unit) =
-        CompositionLocalProvider(
-            LocalUIViewController provides rootViewController,
-            LocalLayerContainer provides rootViewController.view,
-            LocalInterfaceOrientation provides interfaceOrientationState.value,
-            LocalSystemTheme provides systemThemeState.value,
-            content = content
-        )
-
-    @OptIn(ExperimentalComposeApi::class)
-    fun createRootComposeSceneBridge(): ComposeSceneBridge =
-        if (configuration.platformLayers) {
-            createSingleLayerComposeSceneBridge()
-        } else {
-            createMultiLayerComposeSceneBridge()
-        }
-
-    val keyboardVisibilityListener = object : KeyboardVisibilityListener {
-        override fun keyboardWillShow(arg: NSNotification) = composeSceneBridges.forEach {
-            it.keyboardVisibilityListener.keyboardWillShow(arg)
-        }
-
-        override fun keyboardWillHide(arg: NSNotification) = composeSceneBridges.forEach {
-            it.keyboardVisibilityListener.keyboardWillHide(arg)
-        }
-    }
-
-    override val rootViewController: RootUIViewController by lazy {
-        RootUIViewController(
-            configuration = configuration,
-            content = content,
-            createRootComposeBridge = ::createRootComposeSceneBridge,
-            keyboardVisibilityListener = keyboardVisibilityListener,
-            composeSceneBridges = composeSceneBridges,
-            interfaceOrientationState = interfaceOrientationState,
-            systemThemeState = systemThemeState,
-            onViewSafeAreaInsetsDidChange = {
-                composeSceneBridges.fastForEach {
-                    it.updateSafeArea()
-                }
-            }
-        )
-    }
-
-}
-
-private fun getLayoutDirection() =
-    when (UIApplication.sharedApplication().userInterfaceLayoutDirection) {
-        UIUserInterfaceLayoutDirection.UIUserInterfaceLayoutDirectionRightToLeft -> LayoutDirection.Rtl
-        else -> LayoutDirection.Ltr
-    }
